Add `pathway_analysis.R` for DEG pathway analysis
The `pathway_analysis.R` script analyses specified pathways in a DEG
list and provides an image of that pathway with fold changes overlayed
on top. It can also list and quantify the types of interactions and
perturbation events (A -> B is an event if A is a DEG) in said DEG list.